fix: change print type

+++ src/rinterface_extra.c
@@ -73,8 +73,8 @@ enum igraph_versions {
 void R_check_int_scalar(SEXP value)
 {
   if (Rf_xlength(value) != 1) {
-    igraph_errorf("Expecting a scalar integer but received a vector of length %zu.",
-                    __FILE__, __LINE__, IGRAPH_EINVAL, (size_t) Rf_xlength(value));
+    igraph_errorf("Expecting a scalar integer but received a vector of length %d.",
+                    __FILE__, __LINE__, IGRAPH_EINVAL, Rf_xlength(value));
   }
   if (((igraph_integer_t) REAL(value)[0]) != REAL(value)[0]) {
     igraph_errorf("The value %.17g is not representable as an integer.",
@@ -85,16 +85,16 @@ void R_check_int_scalar(SEXP value)
 void R_check_real_scalar(SEXP value)
 {
   if (Rf_xlength(value) != 1) {
-    igraph_errorf("Expecting a scalar real but received a vector of length %zu.",
-                    __FILE__, __LINE__, IGRAPH_EINVAL, (size_t) Rf_xlength(value));
+    igraph_errorf("Expecting a scalar real but received a vector of length %d.",
+                    __FILE__, __LINE__, IGRAPH_EINVAL, Rf_xlength(value));
   }
 }
 
 void R_check_bool_scalar(SEXP value)
 {
   if (Rf_xlength(value) != 1) {
-    igraph_errorf("Expecting a scalar logical but received a vector of length %zu.",
-                    __FILE__, __LINE__, IGRAPH_EINVAL, (size_t) Rf_xlength(value));
+    igraph_errorf("Expecting a scalar logical but received a vector of length %d.",
+                    __FILE__, __LINE__, IGRAPH_EINVAL, Rf_xlength(value));
   }
 }
